improved the ui and fixed a bug
changed the ui, so it shows more data.
Now you can see:
- average response time
- sending connections (as int)
- sending connections (as progress bar)

also fixed the but that to many connections opened. It was caused by the time it takes to create a new TCP Stream. Meanwhile the main loop already created a new thread that also spawned a new connection, since the connection counter did not increase yet.
My solution is just another counter that counts the threads.